// src/common/enums/post.enum.ts
export enum PostStatus {
  DRAFT = "draft",
  PUBLISHED = "published",
  ARCHIVED = "archived",
  DELETED = "deleted",
  REJECTED = "rejected",
  APPROVED = "approved",
  PENDING = "pending",
}

export enum PostVisibility {
  PUBLIC = "public",
  PRIVATE = "private",
}

export enum PostDeleted {
  NO = "no",
  YES = "yes",
}