// src/common/enums/comment.enum.ts
export enum CommentStatus {
  PENDING = "pending",
  APPROVED = "approved",
  REJECTED = "rejected",
}

export enum CommentDeleted {
  NO = "no",
  YES = "yes",
}

export enum CommentVisibility {
  PUBLIC = "public",
  PRIVATE = "private",
}

export enum CommentType {
  TEXT = "text",
  IMAGE = "image",
  VIDEO = "video",
}